Updates RandomizedBenchmarkingPlot to use new RB code.
The new plot mimics the matplotly plot created by the RBResults.plot()
function using plotly, and may replace it in the future. (Note:
RBResults.plot() doesn't import matplotlib at the top level).  This
commit also updates the plotly-to-matplotlib functionality so that
1) it now works with plotly v3 and 2) it converts plotly "box"
traces to matplotlib violin plots.  Creation and pdf-saving of
the new plot has been added to a unit test (in test_rb.py).