Fixes dill pickle bug/issue and updates comparison datasets & gatesets.
There was an issue, only seen in TravisCI runs, whereby DataSets
could not be unpickled because they couldn't import the "dill" module.
This seemed strange in that pyGSTi doesn't use dill to serialize
anything, but upon closer inspection it was (in Python v2.7) using
dill inadvertently because
 1) cvxpy imports dill somewhere,
 2) when dill is imported it sneakily alters the behavior of pickle!,
 3) pyGSTi imports cvxpy (or at least tries to) when it is imported.

What this meant is that all the pickle files (datasets in particular)
which pickled new-style objects contained reference to "dill" (and
could be detected by grepping for "dill") - including those datasets
stored in this repo for unit test comparisons.

This commit fixes this issue by doing the following:
- imports to cvxpy are surrounded by pickle-saving code that
  preserves the original behavior of pickle (at least I think it does).
- all the pickle files (which ended up being just datasets) in the
  cmp_chk_files/ directory containing dill references have been
  updated (regenerated using the new code).  This caused the need to
  also update a number of "expected solution" gateset files which
  needed to change because the datasets did.

Hopefully this fixes the "dill pickle" issues on TravisCI.