Follow up to prior commit: have travisTests.py print to screen.
One reason we needed the "travis_wait" before calling travisTests.py
was that travisTests.py logs all of its output to a file and then
dumps it to the screen - but only after it's all done.  So lots of
TravisCI jobs timed out after removing the "travis_wait" b/c any
*set* of tests that exceeded 10min would be killed.  This commit
has travisTests.py instead just dump its output to the screen (we
don't the need the output saved anyway).