Adds timing information to TravisCI tests.
The test driver runTests.py now accepts the --with-timer or -t
flag to use the nose-timer plugin.  The "travisci" extra requirements
tag now installs nose-timer and travisTests.py enable the timer
for TravisCI tests.  This is just a convenience: now we'll know how
long each of the tests takes in TravisCI jobs.