Updates testCalcMethods1Q.py to gauge opt before comparing gatesets.
Noting that python3.5 vs python3.6 can give different GST estimates
(due to gauge freedoms), this commit updates a couple of the tests
to gauge-opt to the "accepted correct" gateset before comparing to
it.

NOTE: This is not possible in term-based cases where we currently
compare the paramvecs, so thse may break in the future as the gauge
degrees of freedom allow different paramvecs.