Properly raises NotImplementedError in Lindblad gate map todense(...)
When a LindbladParameterizedGateMap uses term-based calculation (has
 evotype == 'svterm' or 'cterm') then todense(...) doesn't (and probably
shouldn't) work but no exception was thrown.  This was an issue as
todense is called from GateSet.frobeninusdist when checking whether
the GateSets on different MPI ranks are equal (i.e. in sync).  Now
this case throws a NotImplementedError, which signals to the checking
code that frobeniusdist isn't implemented and the check doesn't need
to be performed (what we want).