Updates customlm.py with "large-J" and "max-dx" potential enhancements.
Logic attempting to recover from "stepping on a landmine" behavior
where the LM algorithm unknowingly moves to a location with a huge
|J| is added but commented out.  The implemented behavior, which
backtracks if a huge |J| is encountered seems like it should help,
but doesn't in practice - so I'm including it in this commit in
case we ever want to revisit it (it will likely be removed in
a future commit).

The second enhancement does seem to actually help: it prevents the
optimizer from taking any huge steps, i.e. it limits |dx|.  This
can occur when beginning from a singular/special point where |J|
is low (near zero?) but the curvature is high, so the estimated
step is much too large and takes us into bad place where subsequent
steps cannot recover.  Currently the maximum |dx|^2 is 1.0*len(dx),
but we should make the prefactor into a proper argument in the future
(as we should the number of finite-difference steps - still a TODO item).