Updates SPAMVecs & POVMs, fleshing out classes and fixing bugs.
In addition to minor updates & bug fixes, this commit adds significant
functionality to the computational-basis classes:
ComputationalSPAMVec and ComputationalBasisPOVM.  These classes now
support any evotype, eliminating StabilizerZPOVM and StabilizerEffectVec
as these are just evotype=='stabilizer' specializations of the
updated classes.

This commit also updates PureStateSPAMVec to handle it's evotype
more sensibly (it used to just inherit it's pure-state's evotype,
now if the pure-state is "statevec" the PureStateSPAMVec can be
"densitymx", for instance).

Finally, this commit updates LindbladParameterizedSPAMVec to better
separate the SPAM-vec-conversion logic from other portions of the
class.  In particular, from_spamvec_obj will perform conversion
processing (e.g. it can be given a fully-parameterized spamvec as
input) but all other initalization functions expect 1) the "reference"
pure spamvec to have the correct evotype (the same as the evotype of
the LindbladParameterizedSPAMVec being created) and 2) this spamvec
to be *static* - i.e. 0 params - since all the parameters are given
the the error generator (which also needs to have the correct evotype).
Thus, for everything but from_spamvec_obj, the handling of evotypes
and parameters is very simple and nothing is "converted" to something
that isn't passed into the init functions.