Alters LinbladParameterizedPOVM to use a "base POVM".
Now you can construct a LindbladParameterizedPOVM using a base
POVM that contains static (0-parameter) reference spam vecs which
are operated on by the Lindblad error map.  The default is to use
a ComputationalBasisPOVM, but as this only works for qubits, this
additional functionality may prove useful (for qutrits in ptic!).

Text-format GateSet reading/writing has been updated to accommodate
this new functionality.

Because the __init__ signature of LindbladParameterizedPOVM has
changed, we needed to regenerate the JSON gateset reference files
used by testCalcMethods1Q.py.