Updates create_nqubit_sequences to use parallel gate labels & more.
The "& more" is that it also adds the "paramroot" argument that
allows the user to specify what type of parameterization should be
used to construct the sequences (i.e. what parameters need to be
amplified by the sequences).  Perhaps we'll want to make this
even more general in the future where the user can pass in GateSets
directly, but this would be clunkier and may be unnecessary.