Fixes a bug in JSON DataSet serialization.
This commit updates a 2Q JSON dataset file forgotten in the last
commit (now containing the *parallel* gate labels resulting from
create_nqubit_sequences).  In so doing, it was observed that the
DataSet.auxInfo member (a defaultdict) is not serialized properly
to/from JSON - it's read back in as just a *normal* dict.  Now
DataSet.__setstate__ is updated to promote the serialized auxInfo
to a defaultdict if it isn't already one.