Adds relink_parent logic to fix GateSet parameter allocation bug.
This commit fixes the following issue: when a GateSet is serialized
and de-serialized (loaded back), to avoid cycles the 'parent'
member of the GateSet's members is explicitly *not* serialized
(and is replaced w/None - see GateSetChild.__getstate__).  This
means that upon loading, GateSet members have gpindices valid for
the GateSet, but their ._parent == None.  This was delt with in
GateSet.__setstate__ which would set the parent of child objects
via likes like:

for o in self.gates: o._parent = self

This works for *simple* gates (e.g. non-composed/embedded) but
not for gates which contain/depend on other gates or child objects,
like composed and embedded gates, as these dependent objects would
never get their ._parent set (it would still == None).  To remedy
this, this commit adds GateSetMember.relink_parent, which recursively
sets the parent of child objects.  Furthermore, it checks to make
sure that the existing ._parent value is None, as it always should
be after de-serializing.

This commit leaves commented-out some of the print statements used
in debugging.  They are marked in gate.py with "DB:" and "REMOVE".