Fixes a bug in LindbladParameterizedGate.deriv_wrt_params
This bug causes a shape-mismatch error in certain cases, and was
due to deriv_wrt_params returning a cached derivative array instead
of a view.  When subsequent code set the .shape of the array this
altered the cached/stored version within the Gate, which was then
(incorrectly) returned in subsequent calls to deriv_wrt_params.

A similar bug in hessian_wrt_params was fixed in an eariler commit,
but we misse this one back then.