Adds initial create_nqnoise_report for reporting n-qubit results.
Very much a work in progress, this commit copies create_standard_report
to a new create_nqnoise_report function and modifies it and underlying
functions so they (begin to) work with n-qubit (non-dense-gate) gate
set object.  Mainly this has involved removing dense-gate-dependent
stuff from the standard report, but it also add the new
LsGermsSerialStructure object as a way of visualizing the sparse
fiducial pairs of the n-qubit sequences.