Adds writing of auxInfo to text data set files & increases robustness.
Now write_dataset includes "auxInfo" dictionaries as comments when a
DataSet is written.  Secondly, the form of the comments is more
flexible, so that curly brackets are optional (previously they were
not allowed).  So now "Gx 90 10 # 'a': 10" or "Gx 90 10 # {'a': 10}"
both work.  Finally, any warning generated when loading a data set
are accumulated and output at the end rather than as the file is
parsed.  This avoid having them be overwritten (and hence not
really displayed) by the jupyter notebook progress message updates.