Adds better caching and MPI usages to reports.
In particular, for the new n-qubit reports this is useful because
computation of probabilities is expensive.  This update adds
two primary pieces: 1) SmartCache "smartc" arguments to chi2fns,
likelihoodfns, and plothelpers functions so that expensive calls
internal to functions can be cached; 2) comm arguments have been
added to workspace tables and plots so that multiple procs can be
used to compute probabilities, etc.

This commit retains a number of the print statements used to debug
the caching functionality (to ensure cache hit were occurring).
This will be cleaned up in a subsequent commit.