Updates many of the refernce files so unit tests pass.
Due to the change to non-CPTP-constrained "relative" gates in the
n-qubit gate sets (created by build_nqnoise_gateset), a number of
the calcMethods1Q_* nqubit_2Q* and test1Qcalc_* reference files
needed to be updated.  Also, I think the calcMethods1Q_* datasets
needed to be updated so that both python2.7 and python3 versions
are the same (using an earlier machine I think I updated just the
v3 files).  Now most or all of the unit tests should pass using
both python3 and python2.7.