Improves Workspace caching (adds uuids).
This commit improves Workspace caching by using more cache keys
that are persistent across python instances.  The Python default
is that custom objects are hashable, and that hash is based on
the object's id (not persistent).  This commit adds a .uuid
member and custom hashing based on this UUID for GateSet and
GateStringStructure objects.  As many workspace plots/tables
use these objects, this makes caching functions with these
objects in their arguments actually work across filesystem
saves (i.e. saving a cache file and reloading it later).

This commit also changes the internals of SmartCache slightly,
so that it 'digests' each positional argument of a cached function
separately.  Rather than digesting the tuple `(args, kwargs)`, the
'call_key' function now digests `tuple(args) + (kwargs,)`.

Finally, this commit updates report creation functions so they
don't bother the user with a warning when the title is omitted AND
the filename is None (since then no report is actually created).