Updates idle gate (Gi) interpretation/handling for ProcessorSpec.
This commit is meant to at least partially resolve the following
situation:

Multi-qubit circuits typically have a bunch of "whitespace"
which is ideally perfect idle operations (= tensor prod of 1Q
perfect idles).  Perfect-gate GateSets had no real problem,
since they could hold gates like "Gi:2" without any ambiquity.

But noisy GateSets typically would *not* hold imperfect versions
of 1Q perfect idles since a noise 1-qubit idle is not necessarily
still a 1Q gate.  It's more logical to think of simulating circuits
layer-by-layer rather than gate-by-gate, and in this spirit a noisy
GateSet must be able to auto-gate labels like [Gx:1Gy:3] or equivalently
[Gi:0Gx:1Gi:2Gy:3].  Either keeping or dropping the Gi:x labels
seems fine, since it's natural to imply that qubits/lines that aren't
mentioned in a label are idle.  In typical noise models (so far),
there's a single "global" Gi gate (not to be confused w/Gi:x gates)
representing background idle noise and they specific "local noise"
associated with non-idle gate labels (e.g. Gcnot:1:2).  It then
seems wasteful to include in a GateSet gates for Gi:x labels since
these are not even approximately composed, etc., when forming the
n-qubit gate for an entire layer.

An additional issue arises when the Gi:x labels are dropped from
a gate string or circuit in that an entire layer of Gi's gets
mapped to an empty LabelTupTup which wasn't handle-able by the
auto-gators.

The main issues were:
- Gi:x would appear in circuits, but need to be called out specifically
  as something that should be dropped when viewing the circuit as a
  GateString b/c noisy GateSets done contain Gi:x gates.  This seemed
  like a special and unexpected needed Circuit behavior.
- given the above, completely idle circuit layers needed to be represented
  differently (e.g. with a global n-qubit Gi gate) than the rest of the
  layers, since an all-Gi:x layer would be dropped from the corresponding
  GateString entirely.

----------------------------------------------------------------------

The updates in this commit, along with the updated "paradigm" for
n-qubit gatesets is:

- GateSet now has a "auto_idle_gatename" member which holds a gate
  *name* (like "Gi") that can be used within compound (LabelTupTup)
  gate labels to indicate idling qubits.  Such Gi:x gates are *not*
  held in the GateSet, as this would be confusing and inefficient -
  (there's usually just one big "background idle" gate that is used
  as the base for all n-qubit "layer gates".)  This means that the
  GateSets constructed by a ProcessorSpec, which knows what its
  perfect-idle gate name is, are able to handle circuits/strings
  either with or without the Gi:x labels dropped - user's choice.

- GateSet's can also properly process empty LabelTupTups (now
  allowed) due to ComposedGateMap being updated to include a
  separate "dim" argument in its constructor - so you can now
  have a ComposedGateMap that composes 0 factor gates.  This
  means that circuits which do drop the Gi:x labels can have
  an entire layer of Gi:x's map to an empty LabelTupTup and
  everything will work properly (keeping the Gi:x's will work
  too though, as mentioned above).

- Along with the above: Circuits now do *not* drop empty
  layers when converting to GateStrings (tuples) - there
  remains an empty LabelTupTup.

- ProcessorSpec sets the auto_idle_gatename of the models
  (GateSets) it creates, so these GateSets (even the perfect
  "target" ones) don't have Gi:x gates - they're just able to
  process circuits containing these labels.  The philosophy
  moving forward is that any (noise or not) multi-qubit
  GateSets will *not* store perfect or imperfect 1-qubit idle
  gates; they'll use auto_idle_gatename to indicate a
  "whitespace"/"placeholder" name when processing compound
  gate labels (and they might hold a global idle gate - even
  named "Gi" to match 1Q/2Q GST separately).

- Note: the point of giving a Circuit a given idle gatename
  is less important now -- doing so is significant if you're
  constructing circuits w/"whitespace" (so the whitespace
  gets filled in appropriately) *and* don't want the whitespace
  removed at some later stage.  The default .identity is still
  left as 'I' for now.

- Related update: new GateSet.set_state_space_labels(...)
  allows the user to set a GateSet's state space labels.

- RB simulator code, which *does* rely on Circuits w/Gi:x's in them
  has been patched to check for a GateSet's auto_idle_gatename and
  add appropriate Gi:x elements to the error models created despite
  the fact that the GateSet doesn't have any Gi:x gates.