Updates einsum->tensordot calls *with checks still in place*
As einsum is clearer but tensordot is faster, this commit initiates
a process whereby we'll comment out einsum calls and insert equivalent
(but faster) tensordot, etc., calls beneath.  In this commit, many
replacement sites leave both versions uncommented and assert that the
two are equal.  We'll try to perform tests which check that all the
replacements are correct, but if in the future we suspect a tensordot
replacement might not be correct this will be a good commit to come
back to.  Look for "CHECK EINSUM" to find places where
we check whether the replacment and original are the same.