Beginning of feature-idletomography branch. Adds utility functions.
Adds fidpairs_to_gatename_fidpair_list and gatename_fidpair_list_to_fidpairs
functions which convert fiducials pairs in the usual 2-tuple of GateStrings
to the gatename-fidpair-list format (a list of the gate names that get
applied to each qubit; relies on fiducials being composed of just 1Q
gates).  Also fixes create_nqubit_sequences so that when idleOnly=True
it returns a LsGermsSerialStructure as it should, instead of the
list of sequence tuples used previously.