Updates shared-idle auto-gator to work w/more GateSets.
When build_nqnoise_gateset is given addIdleNoiseToAllGates=False
it creates (non-idle) gates with only 2 factors instead of 3 (it
leaves out the global idle factor).  Previously, the shared-idle
auto gator relied on and asserted that each gate have 3 factors,
and so it would fail when addIdleNoiseToAllGates=False gatesets
were used.  This commit fixes this issue, and adds support for the
2-factor gates.