Adds "clip" sample error type to generate_fake_data.
We needed to remove the usual [0,1] probability clipping in
generate_fake_data in order to properly debug affine errors in
idle tomography (needed to allow > 1 and < 0 "counts").  This
commit makes this possible by chaning the "none" sampleError
value to *not* clip the probabilities and adding a new "clip"
value that *does* clip them (just as "none" hade done previously).