Improves the fitOrder=2 idle tomography fit to get derivative @ y=0.
When we fit idle tomography data to a quadratic, the slope we desire
is the slope of the quadratic where it intersects the error-probability
(y) == 0 (if at all, it does this at two points; we want the lower/
leftmost one).  Previously we just took the derivative at x=0, which
is not quite correct.  When there's no SPAM error the quadratics get
pinned to the origin (x=0,y=0) and so this update doesn't matter -
this is relevant for 2nd order fits when there're SPAM errors.