Adds ability to show predicted/true rates in idle tomography reports.
This commit adds the 'predicted_obs_rates' member to the
IdleTomographyResults object, which holds a dictionary w/keys
'hamiltonian', 'stochastic', and/or 'stochastic/affine' to match
those in the results object's pauli_fidpairs member, and after
two more levels of nested dictionaries (one for the pauli fidpair
and another for the observable or outcome) holds the predicted rates.

The `predicted_observable_rates` function in idletomography/tools.py
now creates the type of nested dictionary structure expected to be
in results.predicted_obs_rates['stochastic/affine'], for instance.

Finally, the observable rate workspace plot is updated to show the
predicted rate when it's available via a result object's
predicted_obs_rates member.