Makes GST reports more flexible - to work with "new-style" GateSets.
Lots of ".todense()" additions so that GateSets without native dense
gates can still generate GST reports.  Also, lots of fixing branch
conditions for whether something is a gate-label: previously we just
checked if it was a string, now we also check if it's a Label object.

Also, the base-class deriv_wrt_params (called for map-type gates) for
gates now throws a NotImplementedError rather than an AssertionError
when the gate has a nonzero number of parameters -- this causes this
to be caught by num-nongauge-param code properly (so we fallback to
 just the #params).

Generating reports from "new-style" (using Gx:0-like labels and map-
type gate objects) hasn't been tested extensively, but the few cases
that have been tested work now.