Revamps gatetools.project_to_target_eigenspace(...)
Our algorithm for projection of a gate onto a target gate's
eigenspace has been substantially improved.  Previously, we dealt
wtih the fact that the target gate often has degenerate eigenvalues
in an ad-hoc way (by adding a little of the actual gate matrix to
the target gate matrix), which isn't always reliable.  This commit
adds logic to match gate and target eigenvalues using an improved
min-weight matching that will preserve (if possible) eigenvalue
conjugacy relations (matrixtools.minweight_match_realmxeigs); then
the eigenvectors used for the target gate are chosen by projecting
the actual gate's eigenvectors onto the appropriate eigenspace of
the target gate.

Lots of debug statements remain, to be removed in the next commit.