Adds more automation to Idle Tomography in GST reports.
New determine_paulidicts(...) function in idtcore.py automatically
constructs the "Pauli-basis dictionaries" needed by idle tomography
from a target gateset.  Currently this routine only works for gatesets
containing X and Y pi/2 rotation gates, but it could be expanded in
the future.

Secondly, do_idle_tomography now accepts "auto" as a value for its
extract_<X> arguments, whereby it checks the rank of its Jacobian
matrices and only returns (in a results object) intrinsic values
derived from a full-rank Jacobian.

Both of these updates are now used by create_standard_report so
that users don't need to specify a set of Pauli basis dicts to
get idle tomography to work, and the idle tomography results will
be tailored (e.g. whether affine errors are included or not)
according to how much the GST sequences probe the idle gate.

Finally, IdleTomographyIntrinsicErrorsTable is updated so that
only columns for intrinsic-rate types that are *present* in the
provided IdleTomographyResults object are displayed (even if
others are requested by the 'display' argument).