Adds new IdleTomographyObservedRatesTable report table.
The class formerly named IdleTomographyObservedRatesTable has
been renamed to IdleTomographyObservedRatesForIntrinsicRateTable,
as this shows the observed rates relevant to a single intrinsic
rate (at one time).  The new, seemingly better table shows all
the large observed error rates along with an expression for how
they relate to all the intrinsic rates.  We'll keep the old
IdleTomographyObservedRatesForIntrinsicRateTable around for now
in case we find a use for it later, but both the idle tomography
and GST reports have been updated to use the new table
(IdleTomographyObservedRatesTable) and therefore dispense with
the intrinsic-error-selection switchboard previously found on the
Idle Tomography report tab.

A number of old sections were left commented out and tagged
with "TODO REMOVE" comments.