Changes ham_tmpl parameter to default to "auto" in idle tomog. routines.
Previously the default ham_tmpl parameter was the "standard" set
used for maxweight=2 errors.  Now this is parameters can be (and
defaults to) "auto", which internally uses the same maxweight=2
set and the trivial ("X","Y","Z") set when maxweight=1.  This
essentially makes it unnecessary to specify ham_tmpl when doing
idle tomography with maxweight=1, which is convenient.