Adds idle tomography unit-tests & fixes in preparation for merge.
This will be the final commit on the feature-idletomography branch
before merging into develop.  This commit adds several unit tests
for the idle tomography subpackage, and fixes some issues discovered
in via the tests.  The biggest was that a dict was used where an
OrderedDict should have been in idtcore.py, which only showed itself
in python 2.7 (since as of python3.7 all dicts are insertion-ordered).