Adds proper initialization of subTree.p_polys in TermEvalTree
For tree splitting to work properly with term-based calcs, the
subTree.p_polys, subTree.dp_polys, and subTree.hp_polys need to be
initialized to empty dictionaries instead of None in TermEvalTree's
create_subtrees(...) method.  This is particularly relevant for
use with MPI (but this bug didn't show up in non-MPI contexts).