Fixes and improves GateSet paramvec rebuilding/maintenance.
Messing around with ways of creating GateSets with composed &
embedded gates revealed several issues with the way we were
building and maintaining a GateSet's paramvec based on its contents.

First, this commit fixes a bug in _rebuild_paramvec whereby the
`off` variable (an offset to the point where the next new parameter
should be inserted) was being reset to 0 instead of being left along
when encountering objects with no (zero) parameters.  Fixed now.

Secondly, there was an issue with "dangling objects", e.g. sub-parts
of gates or gates themselves, which retain seemingly valid references
to a GateSet's parameters when they actually don't.  For instance,
when a Gate with 2 parameters and gpindices = [0,1], say for key
 ('Gx',0), is replaced with another Gate with zero parameters, the
_rebuild_paramvec function will dutifully remove the now unreferenced
parameters 0 & 1 from the GateSet -- but the removed gate will *still*
have its parent set to the GateSet and have its gpindices == [0,1]!
This is fine as long as the removed Gate isn't used to construct
future gates, but if it is then it will errantly point to
parameters that either don't exist or that are not the parameters
it should be pointing to.

To fix this, we now try to reset to None the .parent member of
gates (and their sub-components) which are removed from a GateSet.
This is initiated in labeldicts.py which calls (the new)
`unlink_parent` on any object that is being replaced in a GateSet.
This gives the object a chance to check with its parent to see if
the parent contains any more references to the object (as it might
if the object is a sub-component of several GateSet members) and set
its parent to None if the reference count is zero.  This reference
count checking is done via the new `_obj_refcount` methods of
GateSet and GateSetMember, which perform checking in sub-components
too.  The upshot of this is that any Gate or sub-component thereof
will (if this machinery works properly) have its .parent reset to
None (thereby invaliding its .gpindices) whenever its parent
GateSet no longer contains any references to it.