Fixes a bug in GateSet._embedGate (auto-embedding)
Overzealous short-circuiting of embedding resulted in gates acting
on the full Hilbert space but with a *different* sector ordering
*not* being wrapped in an EmbeddedGate when they should have been.
For example, Gcnot:1:0 in a 2-qubit system still needs to be
wrapped in an embedded gate to reorder the state space sectors
0,1 -> 1,0.  Previously the ordering was ignored and a gate was
not embedded if whenever it acted on the full state space.  Fixed
now by additional conditions in GateSet._embedGate (used for auto-
embedding).