Adds ComposedErrgen and factors out "submembers" processing.
As part of an ongoing task of making error generators their own
operators (i.e. separating them from the Lindblad gates), this
commit adds ComposedErrgen, which represents the sum of one (zero?)
or more error generators.  This makes it similar to ComposedGateMap,
except it composes by adding rather than matrix-multipying/applying
and must present the still-being-determined interface for an
"error generator" object.  This commit contains the initial and
completely untested version of ComposedErrgen.

A note for future development - we may want to subclass ComposedErrgen
for the specific case when it's held as a dense matrix, similarly to
ComposedGateMap and ComposedGate now... not sure how useful this
would be.  Currently, methods like deriv_wrt_params only work when
the error generator is non-sparse; maybe this is good enough.

More generally, this commit consolidates the set_gpindices,
allocate_gpindices, relink_parent, and unlink_parent methods in
many of the Gate/SPAMVec classes which are essentially wrappers
for sub-objects (at least in terms of what parameters they have).
We introduce the notion of "submembers", identified by the new
GateSetMember.submembers() function which defaults to nothing ([])
but derived classes can override.  This allows GateSetMember
methods like set_gpindices, unlink_parent, etc., do the common
processing required when a class has sub-members.  In the end,
this means the derived classes only need to implement submembers()
and copy() [the latter because there's no easy way for the base
GateSetMember class to know how to construct a copy of a derived
object based on copies of the submembers (different __init__
signatures)].  Maybe we can find a clever way to get a base class
copy() function in the future.

A final minor note: we introduce the concept of a "null slice" --
i.e. a slice with zero elements -- and pick slice(0,0,None) to
represent this.  Functions in gatesetmember.py now treat null slices
specially, since it doesn't make sense to require that a sub-member
with 0 parameters (like a StaticGate in an EmbeddedGate) have its
gpindices equal slice(6,6,None) just because the EmbeddedGate has
gpindices equal to slice(0,6,None) or slice(6,12,None), etc. -- we
just want the static gate to have gpindices equal to slice(0,0,None)
and that should be fine with everything else.