Adds EmbeddedErrorgen class (completely untested).
This rounds out the classes that we'll need to flexibly represent
error generators as their own objects.  Like ComposedErrorgen,
the new EmbeddedErrorgen is completely untested and likely to
contain bugs, but the right ideas for the implementation (except
for the derivative & hessian functions, which I've punted on so far)
are there.  For better organization and clarity, having a base
class ("ErrorGenerator" ?) would be helpful so the interface
error generators are expected to implement is well-defined.