Fixes bugs related to new Lindblad/Errorgen classes (found by unittests).
The big issue here is that copying Lindblad gates wasn't working
because a new (and temporary) "full_unitary_postfactor" member of
LindbladParameterizedGateMap was used in .copy().  This fixes that
issue by removing the temporary member, and also fixing the call
of GateMatrix.__init__(...) in LindbladParameterizedGate.__init__
to correctly use the previously constructed self.base rather than
re-initializing (the post factor for this gate) with the identity
matrix.