Adds 'errcomp_type' to build_nqnoise_gateset(...).
New 'errcomp_type' argument specifies how errors are composed:
"gates" composes the gates/ops corresponding to different errors,
whereas "errorgens" uses the new error generator classes to embed
and compose error generators before ultimately creating a single
"error gate" that is exp(sum_of_errorgens).  This commit adds
the framework for creating such gates, and 'errcomp_type' replaces
all 'loc_noise_type' occurences in nqnoiseconstruction.py (and more),
which used to give options for making "manylittle" or "onebig"
Lindblad term (the "manylittle" is essentially the new "gates"
option and "onebig" is obsolete given the new error generator classes
and so has become "errorgens" but is implemented much then "onebig"
was).

GST with a "errorgens" GateSet has been tested and works except
that it's dog slow due to needing to call EmbeddedErrorgen.tosparse()
(which is really just EmbeddedGateMap.tosparse()) so frequenctly -
we'll need to give the error generators proper representations and
implement "torep" for them in the future to make this usable.

Also, the SharedIdleAutogator has been given a 'errcomp_type'
argument but currently does nothing with it - so using parallel
gates will break on a "errorgens" GateSet until this implementation
is finished.