Updates VerbosityPrinter to have `split` option & minor improvements.
This commit adds a 'split' argument to the constructor of a
VerbosityPrinter, which when True causes the printer to send
output to both stdout and a specified file.

This commit also removes an extra '\n' in _put when output is
sent to file, which was causing file output to be double-spaced.

Other than this, there are a couple minor changes carrying over
from fixes Lucas made.