Major update & revamp to Circuit object.  No GateString/OpString needed.
This commit updates the Circuit object so internally it holds a
tuple (when "static" or non-editable) or list (when editable) of
labels rather than a list-of-lists grid with duplicated labels as
it had previously.  It reproduces most/all of the old Circuit
functionality and adds more intuitive indexing & assignment
capabilities (which should be described in a Tutorial later).

A static Circuit now functions as a GateString used to (essentially
as a tuple of layer-Labels) and so there's no need for the latter.
While opstring.py still exists as of this commit, it has no
dependencies and wil be removed in the near future.

All of the pyGSTi code (including unit tests) have been updated
to use the new Circuit, and all unit tests now pass except for
drift-related tests that are known to fail and, more importantly,
*the loading of old pickled objects*.  This will likely be
remedied in the future but as of now:

DON'T PULL HERE if you need to work with previously pickled
objects.

Finally, note that the unit tests till create new reference
files and these reference files are *not* updated by this
commit.