Revamps Model (old GateSet) class to Model, ExpicitOpModel, ImplicitOpModel.
This commit splits what was the GateSet class (the Model class
after the recent massive-rename) into two classes - ExplicitOpModel,
which contains operators (preps,povms,and operations) for *layers*
and can only simulate layers for which it has an explicit element
in one of it's member dicts (e.g. .preps, .povms, .operations,
.instruments).  The ImplicitOpModel class instead creates layer
operations on the fly via a LayerLizard (former ~AutoGator) class
from "building block" elements stored in the model's .prep_blks,
.povm_blks, .operation_blks, .instrument_blks members.  A new
ExplicitOpModel_Calc class holds the non-essential calculation
routines (formerly in GateCalc) which are not needed (or easily
implemented) for the implicit models.

Thus, this commit updates build_nqnoise_model and build_nqubit_model
functions to return ImplicitOpModels and associated with each
function is a LayerLizzard class that is passed to the ImplicitOpModel
constructor and supplies the logic for on-the-fly layer creation.

Tests are also updated so all tests pass except known-failing
drift tests and tests for loading old files.