Updates unit tests given recent renamings/fixes.
Tests are updated so all tests pass except known-failing
drift tests and tests for loading old files.  The unit tests
still create new reference files and these reference files and
so the reference files in the repo at this point remain stale and
in need of future updating (the ref files are *not* updated by this
commit).