Performs one (maybe final?) round of minor renaming.
Renames variables and functions, particularly in the reportables
and report generation code, to eliminate the word "gateset" and
"gatestring", as these are no longer objects in pyGSTi (now they're
"model" and "circuit").  Replacements performed in unit test also.

The one remaining area that hasn't been renamed yet is the HTML
report templates, where we still use "gateset" and "gatestring"
in the key names for insertion into template files (and so these
names appear in factory.py too).