Fixes some bugs related to creating reports with models in a custom Basis.
When computing error generator projections there seems to have been
a past reason to only convey the '.name' of a Model's matrix (operation)
Basis instead of the entire basis object itself.  Comments indicate that
the projection basis isn't the same dimension as the matrix basis, but
I don't think this is correct.  This may have been thought to be an issue
because the ProjectionsBoxPlot code would attempt to compute new Basis
objects of desired dimension from the name of the Basis it is passed (e.g.
a 2-qubit ProjectionsBoxPlot will try to construct 1Q (dim=2) Basis
objects to construct the x- and y-axis labels) - I think this is why it
seemed the projection basis "needed" to be a different dimension.

Passing the name rather than the full Basis object obviously causes
problems when one has constructed a custom basis which cannot be
simply re-generated whenever it's needed from it's name.

This commit hopefully fixes this issue by:

1) passing Basis objects rather than just their names within the
projection-plot code (in workspacetables.py and reportables.py).
This meant lots of basis.name -> basis and mxBasis.name -> mxBasis.

2) updating ProjectionsBoxPlot to gracefully handle the case when
it's given a Basis object and/or cannot construct a Basis it needs
for axis labels (it just leaves the labels blank as a last resort).