Begins tutorial/example notebook restructuring.  New skeleton.
Creates a new organization for the tutoral and example notebooks,
adding a single top-level START_HERE.ipynb and two high-level
overview notebooks in Tutorials/ .  The previous tutorials have
been largely renamed and placed within directories beneath Tutorials/.
The idea is that the high-level-overview tutorials will link to
other tutorials to more efficently guide users to relevant
information.

Note: while the basic structure is in place, there are a number of
blank placeholder notebooks, and the FAQ and examples are at this
point broken and in need of revision/relinking.