Fixes issues with DataSet.get_degrees_of_freedom()
Adds `method` (="all_outcomes-1") and `aggregate_times` (=True)
arguments to allow other behavior, but by default we not just
could the total number of possible outcomes on each line, regardless
of the number of distinct times, as the degrees of freedom for that
line (row).  This should fix some issues with running GST on time-
series data, but it may raise issues with multi-qubit GST where
there's going to be huge numbers of DataSet degrees of freedom if
we cound them in this way.